+++ lib/slim_lint/linter/comment_control_statement.rb
@@ -4,6 +4,7 @@ module SlimLint
   # Searches for control statements with only comments.
   class Linter::CommentControlStatement < Linter
     include LinterRegistry
+    support_autocorrect
 
     on [:slim, :control] do |sexp|
       _, _, code = sexp
@@ -16,7 +17,10 @@ module SlimLint
 
       report_lint(sexp,
                   "Slim code comments (`/#{comment}`) are preferred over " \
-                  "control statement comments (`-##{comment}`)")
+                  "control statement comments (`-##{comment}`)",
+                  correction: ->(line) {
+                    line.sub(/-\s*#{Regexp.escape(code.to_s)}/, "/#{comment}")
+                  })
     end
   end
 end