+++ lib/slim_lint/linter/empty_control_statement.rb
@@ -4,12 +4,14 @@ module SlimLint
   # Searches for control statements with no code.
   class Linter::EmptyControlStatement < Linter
     include LinterRegistry
+    support_autocorrect
 
     on [:slim, :control] do |sexp|
       _, _, code = sexp
       next unless code[/\A\s*\Z/]
 
-      report_lint(sexp, 'Empty control statement can be removed')
+      report_lint(sexp, 'Empty control statement can be removed',
+                  correction: ->(_line) { nil })
     end
   end
 end